More Details Added
Bitcoin is an experimental digital currency that enables instant payments to
anyone, anywhere in the world. Bitcoin uses peer-to-peer technology to operate
with no central authority: managing transactions and issuing money are carried
out collectively by the network. Bitcoin Core is the name of open source
software which enables the use of this currency. "The price of Bitcoin can vary,
due to its worth in the market between investors." (Quote is the only text edited from original, adding more important introduction to the already simplified details.)